Replace SessionFlags with bool to open session
There are only two flags supported as arguments when opening a
session.

One of them must always be true, but perversely defaults to false.
This forces client code to construct a trivial value to pass. This
commit now hides this flag, setting it to its only valid value always.
This also removes a single test which checked for failure when the flag
was set to false.

With only one flag (read/write) remaining, the open session call now
accepts a boolean for the option and conversion to a wider integer
type is handled internally.

Signed-off-by: Keith Koskie <vkkoskie@gmail.com>