Remove InitializeFlags
This is only two flags, one of which isn't supported yet, and the
structure itself was effectively private despite having pub
visibility.  Because the initialization APIs need considerable
attention outside the scope of this branch, the change avoids
any elaborate redesign and just falls back to using raw CK_FLAGS.

Because this is the last use of the Flags trait, it is also removed.

Signed-off-by: Keith Koskie <vkkoskie@gmail.com>