Fix codespell action
Previously the action did not checkout the repository and as such
actually didn't do what it was supposed to.

This patch fixes all typos that accumulated over the time. Adds the
checkout action and fixes the codespell version action to `v1`.

Additionally codespell settings have been moved to a file so that the
`codespell` command line can be executed locally using the same
settings as in the CI.

Signed-off-by: Wiktor Kwapisiewicz <wiktor@metacode.biz>